# openapi.yml
openapi: 3.0.0
info:
  title: Botapeer API
  description: Botapeer API
  version: 1.0.0
servers:
  - url: "{server}/api"
    variables:
      server:
        default: http://localhost:8081
        enum:
          - http://localhost:8081
          - https://api.botapeer.com
paths:
...
  '/posts/plant_records/{plantRecordId}':
    post:
      operationId: createPost
      tags:
        - Post
      description: 投稿作成
      parameters:
        - in: path
          name: plantRecordId
          description: 生育記録ID
          required: true
          schema:
            type: string
            example: 1
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              properties:
                formData:
                  $ref: '#/components/schemas/CreatePostFormData'
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 401(Unauthorized)
          headers:
            http_status:
              description: HTTPステータス
              schema:
                type: integer
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'          
  '/posts/{postId}':
    get:
      operationId: getPostById
      tags:
        - Post
      description: 任意の投稿をpostIdから取得
      parameters:
        - in: path
          name: postId
          description: 投稿ID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/PostResponse'
        '404':
          description: 404(User Not Found)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deletePost
      tags:
        - Post
      description: 任意の投稿をpostIdから削除
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: 投稿ID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: ok
        '404':
          description: 404(Post Not Found)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'  
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/ErrorResponse'
  '/posts/{postId}/users/{userId}/likes':
    post:
      operationId: createLikeToPost
      tags:
        - Post
      description: 投稿記事のLike作成
      parameters:
        - in: path
          name: postId
          description: 投稿ID
          required: true
          schema:
            type: string
            example: 1
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 401(Unauthorized)
          headers:
            http_status:
              description: HTTPステータス
              schema:
                type: integer
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteLikeToPost
      tags:
        - Post
      description: 投稿記事のLike削除
      parameters:
        - in: path
          name: postId
          description: 投稿ID
          required: true
          schema:
            type: string
            example: 1
        - in: path
          name: userId
          description: ユーザーID
          required: true
          schema:
            type: string
            example: 1
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 401(Unauthorized)
          headers:
            http_status:
              description: HTTPステータス
              schema:
                type: integer
        '405':
          description: 405(Validation exception)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                
...
components:
  schemas:
...
    PostResponse:
      properties:
          id:
            type: integer
            format: int64
          plantRecordId:
            type: integer
            format: int64
          title:
            type: string
          article:
            type: string
          imageUrl:
            type: string
          status:
            type: integer
          createdAt:
            type: string
            format: date-time
          updatedAt:
            type: string
            format: date-time
          like:
              $ref: '#/components/schemas/LikeResponse'
    LikeResponse:
      properties:
        count:
          type: integer
          format: int64
        isLikeWithRequestUser:
          type: boolean
...
    CreatePostFormData:
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 40
        article:
          type: string
          minLength: 1
          maxLength: 10000
      required:
        - title
        - article
    ErrorResponse:
      properties:
        errors:
          $ref: '#/components/schemas/Error'
    Error:
        type: array
        items:
          properties:
            code:
              type: integer
            message:
              type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT